Dev test file
Created the file to test certain functions, NOT PRODUCTION READY USE AT YOUR OWN RISK.

Takes the highest prices in a block and averages them to see if the price margin is good. Also takes the latest chargingprice and verifies it against a certain price to know if it's good to discharge.

When price gets under the profit margin the reserve mode is put on the current SOC and the switch is pulled on.